added hometasks 4

+++ src/lesson4/terminal/ClientOfBank.java
@@ -3,9 +3,9 @@ package lesson4.terminal;
 public class ClientOfBank {
     private String name;
     private double money;
-    private final int pin;
+    private final byte pin;
 
-    public ClientOfBank(String name, double money, int pin){
+    public ClientOfBank(String name, double money, byte pin){
         this.name = name;
         this.money = money;
         this.pin = pin;
